@@ -7,13 +7,19 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "formik": "^2.4.5",
     "http-server": "^14.1.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-loader-spinner": "^5.4.5",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
     "react-scripts": "4.0.3",
-    "web-vitals": "^1.0.1"
+    "react-transition-group": "^4.4.5",
+    "web-vitals": "^1.0.1",
+    "yup": "^1.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -42,5 +48,4 @@
   "devDependencies": {
     "sass": "^1.38.0"
   }
-
 }
