@@ -1,86 +1,60 @@
-import { Component } from 'react';
+import {useState, useEffect } from 'react';
 
 import './randomChar.scss';
 import mjolnir from '../../resources/img/mjolnir.png';
-import MarvelServices from '../../services/MarvelService';
+import useMarvelServices from '../../services/MarvelService';
 import Spinner from '../widgets/spiner/Spiner';
 import ErrorMessage from '../ErrorMessage/ErrorMessage'
 
-class RandomChar extends Component {
+const RandomChar = () =>{
 
-    state = {
-        char : {},
-        loading: true,
-        error: false
+    const [char, setChar] = useState({})
+    const {loading, error, getCharacter , clearError} = useMarvelServices();
+
+    // eslint-disable-next-line
+    useEffect(() => {
+        updateChar();
+    }, [])
+
+    const onCharLoaded = (char) =>{
+        setChar(char)
     }
 
-
-    marvelService = new MarvelServices();
-
-    componentDidMount(){
-        this.updateChar();
-    }
-
-    componentWillUnmount(){
-        clearInterval(this.interval);
-    }
-
-    onCharLoaded = (char) =>{
-        this.setState({char, loading: false});
-    }
-
-    onCharLoading = () =>{
-        this.setState({
-            loading: true
-        })
-    }
-
-    onError = () =>{
-        this.setState({
-            error: true,
-            loading: false
-        })
-
-    }
-
-    updateChar = () =>{
+    const updateChar = () =>{
+        clearError();
         const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
-        this.onCharLoading();
-        this.marvelService
-        .getCharacter(id)
-        .then(this.onCharLoaded)
-        .catch(this.onError)
+        getCharacter(id)
+        .then(onCharLoaded).catch((e)=> {
+            console.error(e.message);
+        })
+        
     }
-  
 
-    render () {
-        const {char, loading, error} = this.state;
-        const errorMassage = error ? <ErrorMessage/> : null;
-        const spinner = loading ? <Spinner/> : null;
-        const content = !(loading || error) ? <Viev char={char}/> : null;
+    const errorMassage = error ? <ErrorMessage/> : null;
+    const spinner = loading ? <Spinner/> : null;
+    const content = !(loading || error) ? <Viev char={char}/> : null;
 
-        return (
-        <div className="randomchar">
-            
-            {errorMassage}
-            {spinner}
-            {content}
-            <div className="randomchar__static">
-                <p className="randomchar__title">
-                    Random character for today!<br/>
-                    Do you want to get to know him better?
-                </p>
-                <p className="randomchar__title">
-                    Or choose another one
-                </p>
-                <button className="button button__main" onClick={this.updateChar}>
-                    <div className="inner">try it</div>
-                </button>
-                <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
-            </div>
+    return (
+    <div className="randomchar">
+        
+        {errorMassage}
+        {spinner}
+        {content}
+        <div className="randomchar__static">
+            <p className="randomchar__title">
+                Random character for today!<br/>
+                Do you want to get to know him better?
+            </p>
+            <p className="randomchar__title">
+                Or choose another one
+            </p>
+            <button className="button button__main" onClick={updateChar}>
+                <div className="inner">try it</div>
+            </button>
+            <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
         </div>
-        )
-    }
+    </div>
+    )
 }
 
 const Viev = ({char}) => {
